requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'make build' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.
Flask==0.10
simplejson==3.6.5
pytz==2014.9
google-api-python-client==1.3.1
flask-restful==0.2.1
Flask-WTF==0.9.5
requests==2.3.0
urllib3==1.19
